About to extend my primary partition comitting changes so far

+++ MetalLink.Application/Customers/Queries/GetCustomerByIdQuery.cs
@@ -3,4 +3,4 @@ using MetalLink.Shared.Customers;
 
 namespace MetalLink.Application.Customers.Queries;
 
-public sealed record GetCustomerByIdQuery(long CustomerId) : IRequest<CustomerDto?>;
+public sealed record GetCustomerByIdQuery(int CustomerId) : IRequest<CustomerDto?>;
