+++ MetalLink.Shared/Prices/PriceDto.cs
@@ -4,8 +4,8 @@ namespace MetalLink.Shared.Prices;
 
 public class PriceDto
 {
-    public long PriceId { get; set; }
-    public long ProductId { get; set; }
+    public int PriceId { get; set; }
+    public int ProductId { get; set; }
     public decimal PriceA { get; set; }
     public decimal PriceB { get; set; }
     public decimal PriceC { get; set; }
