+++ MetalLink.Shared/Prices/PriceLookupDto.cs
@@ -2,8 +2,8 @@ namespace MetalLink.Shared.Prices;
 
 public class PriceLookupDto
 {
-    public long PriceId { get; set; }
-    public long ProductId { get; set; }
+    public int PriceId { get; set; }
+    public int ProductId { get; set; }
     public decimal PriceA { get; set; }
     public decimal PriceB { get; set; }
     public decimal PriceC { get; set; }
